requirements.txt
python-dotenv
pandas
requests
google-cloud-bigquery
db-dtypes
PyYAML
